refactor: fix coin service messagingTemplate to eventPublisher
